requirements.txt
absl-py==0.15.0
antspyx==0.2.9
appnope==0.1.3
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
blis==0.7.8
cachetools==4.2.4
catalogue==2.0.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
chart-studio==1.1.0
clang==5.0
click==8.0.3
cloudpickle==2.0.0
configparser==5.2.0
crccheck==1.1
cryptography==36.0.1
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
deep-pipe @ git+https://github.com/neuro-ml/deep_pipe@8247580b806d7004a52cfa4d70c8ff1ba56122d4
dicom2nifti==2.3.0
docker-pycreds==0.4.0
executing==0.10.0
fastai==2.7.9
fastai2==0.0.30
fastcore==1.5.21
fastdownload==0.0.7
fastprogress==1.0.3
Flask==2.0.2
flask-paginate==2022.1.8
flatbuffers==1.12
fslpy==3.7.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.25
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==20.0.4
h11==0.13.0
h5py==3.1.0
idna==3.3
imageio==2.10.1
ipython==8.4.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kaggle==1.5.12
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.6
langcodes==3.3.0
lazycon==0.3.1
loky==2.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.6
murmurhash==1.0.8
networkx==2.6.3
nibabel==3.2.1
numpy==1.23.2
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
outcome==1.1.0
packaging @ file:///private/tmp/sip--packaging-20220416-83003-1hqovwb/packaging-21.3
pandas==1.3.4
parso==0.8.3
pathtools==0.1.2
pathy==0.6.2
patsy==0.5.2
pdp==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
ply==3.11
preshed==3.0.7
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
pydicom==2.2.2
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing @ file:///private/tmp/sip--pyparsing-20220416-83003-17i9c2w/pyparsing-3.0.8
PyQt3D==5.15.5
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQt6==6.2.0
PyQt6-3D==6.2.0
PyQt6-Charts==6.2.0
PyQt6-DataVisualization==6.2.0
PyQt6-NetworkAuth==6.2.0
PyQt6-sip==13.1.0
PyQtChart==5.15.5
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.5
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
selenium==4.1.2
sentry-sdk==1.5.1
shortuuid==1.0.8
sip==6.6.1
six==1.15.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.4.0
statsmodels==0.13.0
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-easy==0.2.3
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
text-unidecode==1.3
thinc==8.1.0
threadpoolctl==3.0.0
tifffile==2021.10.12
toml==0.10.2
torch==1.12.1
torchaudio==0.8.0
torchvision==0.9.0
tqdm==4.62.3
traitlets==5.3.0
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.2
typing-extensions==3.7.4.3
urllib3==1.26.7
wandb==0.12.9
wasabi==0.10.1
wcwidth==0.2.5
webcolors==1.11.1
Werkzeug==2.0.2
wrapt==1.12.1
wsproto==1.1.0
wxPython==4.1.1
yaspin==2.1.0
